requirements.txt
# ========================================
# Core dependencies - 核心依赖
# WebSocket and Exchange API
python-binance>=1.0.17
websocket-client>=1.7.0
requests>=2.31.0

# Data Processing (removed as per new structure, but keeping original if not explicitly removed)
# numpy>=1.24.0
# pandas>=2.0.0
# scipy>=1.11.0

# Configuration
pyyaml>=6.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# ========================================
# Logging & Monitoring
# ========================================
colorama>=0.4.6  # Colored console output
python-dateutil>=2.8.2

# ========================================
# Optional: Alert Integrations
# ========================================
# Telegram alerts (uncomment if needed)
# python-telegram-bot>=20.0

# Discord alerts (uncomment if needed)
# discord-webhook>=1.3.0

# ========================================
# Development & Testing
# ========================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# ========================================
# Utilities
# ========================================
python-dotenv>=1.0.0  # For .env file support (API keys)
tabulate>=0.9.0  # Pretty table printing
